refactor: change the project structure
