+++ app/modules/albums/services/album.service.ts
@@ -1,3 +1,3 @@
-import { DataSource } from '../../../DataSource';
+import { DataSource } from '../../common/services/common.services';
 
 export const albumsSource = new DataSource(process.env.ALBUMS_URL as string);
