+++ app/modules/artists/services/artist.service.ts
@@ -1,3 +1,3 @@
-import { DataSource } from '../../../DataSource';
+import { DataSource } from '../../common/services/common.services';
 
 export const artistsSource = new DataSource(process.env.ARTISTS_URL as string);
