+++ app/modules/bands/services/band.service.ts
@@ -1,3 +1,3 @@
-import { DataSource } from '../../../DataSource';
+import { DataSource } from '../../common/services/common.services';
 
 export const bandsSource = new DataSource(process.env.BANDS_URL as string);
