+++ app/modules/genres/services/genre.service.ts
@@ -1,3 +1,3 @@
-import { DataSource } from '../../../DataSource';
+import { DataSource } from '../../common/services/common.services';
 
 export const genresSource = new DataSource(process.env.GENRES_URL as string);
