+++ app/modules/tracks/services/track.service.ts
@@ -1,3 +1,3 @@
-import { DataSource } from '../../../DataSource';
+import { DataSource } from '../../common/services/common.services';
 
 export const tracksSource = new DataSource(process.env.TRACKS_URL as string);
